# src/docs/api.yml
swagger: "2.0"
info:
 description: "Service to Register and provide management for Object identifiers. Extra OID info @ http://www.oid-info.com/"
 version: "1.0.0"
 title: "OID Registry"
host: "localhost:8080"
basePath: "/oidRegistry/v1"
schemes:
- "http"
...
paths:
 /object:
   post:
     tags:
      - OID_NODE
     summary: Register a new OIDNode. Need to recieve full OID
     responses:
       200:
         description: The Object that matches the given OID
       400:
         description: Invalid Request. Need to pass OID when registering a node.
       422:
         description: Cannot Register Node because the OID already exists or the parent node has a Generator associated with it, in which case children nodes must be generated.
...
 /object/{OID}/children:
   parameters:
   - in: path
     name: OID
     type: string
     required: true
     description: The OID for the Node we are retrieving
   get:
     tags:
      - OID_NODE
     summary: Retrieves all Children (Paginated) for the given OID
     responses:
       200:
         description: The Object that matches the given OID
...